refactor: spit `core.py` (#2743)
